# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Khalaf Olive Groves - Ancient Grove Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/bottling-runs/{token}:
    get:
      operationId: getPublicBottlingRun
      tags: [public]
      summary: Sanitized bottle-to-tree traceability dossier looked up by public token (no auth)
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public bottle dossier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicBottleDossier"
        "404": { description: Not found }
  /public/trees/{id}:
    get:
      operationId: getPublicTree
      tags: [public]
      summary: Sanitized public tree profile (no auth)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Public tree profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTreeProfile"
        "404": { description: Not found }
...
components:
  schemas:
...
    PublicBottleDossier:
      type: object
      required: [bottlingRun, sources, treeCount, groveBreakdown, topTrees, labResults, totalContributionKg, totalLitersDrawn]
      description: Sanitized public payload for /bottle/:token. Excludes internal notes and raw lab IDs.
      properties:
        bottlingRun:
          type: object
          required: [runCode, bottledAt, status, singleTree, singleGrove, publicToken]
          properties:
            runCode: { type: string }
            bottledAt: { type: string }
            label: { type: string, nullable: true }
            lotCode: { type: string, nullable: true }
            format: { type: string, nullable: true }
            bottleSizeMl: { type: integer, nullable: true }
            bottlesProduced: { type: integer, nullable: true }
            totalLitersBottled: { type: number, nullable: true }
            singleTree: { type: boolean }
            singleGrove: { type: boolean }
            status: { type: string }
            publicToken: { type: string }
        sources:
          type: array
          description: Per-oil-batch upstream chain (mill, season, pressing delay) for the curious.
          items:
            type: object
            required: [oilBatchCode, litersDrawn]
            properties:
              oilBatchCode: { type: string, nullable: true }
              litersDrawn: { type: number }
              batchCode: { type: string, nullable: true }
              seasonName: { type: string, nullable: true }
              millName: { type: string, nullable: true }
              pressingDelayHours: { type: number, nullable: true }
              harvestDate: { type: string, nullable: true, description: "Harvest batch date (ISO date)." }
              pressingStartedAt: { type: string, nullable: true, description: "Pressing run start (ISO datetime)." }
              pressingCompletedAt: { type: string, nullable: true, description: "Pressing run completion (ISO datetime)." }
        totalContributionKg: { type: number }
        totalLitersDrawn: { type: number }
        treeCount: { type: integer }
        groveBreakdown:
          type: array
          items:
            type: object
            required: [groveName, contributionKg, sharePct, treeCount]
            properties:
              groveName: { type: string, nullable: true }
              groveCode: { type: string, nullable: true }
              contributionKg: { type: number }
              sharePct: { type: number }
              treeCount: { type: integer }
        topTrees:
          type: array
          items:
            type: object
            required: [treeId, treeCode, contributionKg, sharePct]
            properties:
              treeId: { type: integer }
              treeCode: { type: string }
              groveName: { type: string, nullable: true }
              variety: { type: string, nullable: true }
              ancientStatus: { type: string, nullable: true }
              estimatedAgeClass: { type: string, nullable: true }
              crownDiameterM: { type: number, nullable: true }
              contributionKg: { type: number }
              sharePct: { type: number }
              photoUrl: { type: string, nullable: true }
        labResults:
          type: array
          description: Cautious public-facing lab summaries with derived EVOO + health-claim flags.
          items:
            type: object
            properties:
              sampleDate: { type: string, nullable: true }
              labName: { type: string, nullable: true }
              acidity: { type: number, nullable: true }
              peroxideValue: { type: number, nullable: true }
              totalPolyphenolsMgKg: { type: number, nullable: true }
              k232: { type: number, nullable: true }
              k270: { type: number, nullable: true }
              isExtraVirgin: { type: boolean, nullable: true }
              isHealthClaimEligible: { type: boolean, nullable: true }
    PublicTreeProfile:
      type: object
      required: [id, treeCode]
      properties:
        id: { type: integer }
        treeCode: { type: string }
        variety: { type: string, nullable: true }
        ancientStatus: { type: string, nullable: true }
        estimatedAgeClass: { type: string, nullable: true }
        crownDiameterM: { type: number, nullable: true }
        crownAreaM2: { type: number, nullable: true }
        currentHealthIndex: { type: number, nullable: true }
        groveName: { type: string, nullable: true }
        groveCode: { type: string, nullable: true }
        centroidLat: { type: number, nullable: true }
        centroidLon: { type: number, nullable: true }
        photoCount: { type: integer }
        photos:
          type: array
          description: Up to 6 most recent photos for this tree.
          items:
            type: object
            required: [id, fileUrl]
            properties:
              id: { type: integer }
              fileUrl: { type: string }
              thumbnailUrl: { type: string, nullable: true }
              capturedAt: { type: string, nullable: true }

    # ─── Year Report ───────────────────────────────────────────────────────────
...